feat: add skeleton for training pipeline
This commit lays the foundations for implementing a training
pipeline with PyTorch.

An example script is provided in examples/learning, a new folder
created for the purpose of holding sample scripts just for the
machine learning subpackage.

The whole experiment (training) is configured via a JSON file and
a logger for it is also setup with the appropriate JSON file.
Samples for both of them are provided in the examples folder.

Right now, the training pipeline exhibits the most basic functions
and most of the code is to be completed. As of now, it just makes
use of a configuration parser module to parse the JSON file for
the configuration of the experiment and the logger. As a result,
the logger outputs an initialization message.

From now on there are many modules left to implement: trainers,
data loaders, models, layers, losses, and metrics.

See: #9