feat: add loss logic to learning pipeline
Now losses can be created for the learning pipeline. The logic for
building them has been introduced in the training example as well.

The losses module is located in `learning/losses` and it works the
same way as the models. Documentation about that and how to create
new losses has been added to the `README.md` file.