feat: add scheduler and optimizer to training pipeline
Scheduler and optimizers can be created in the training pipeline
in `examples/learning/train.py`. No custom classes nor modules are
used. Unlike losses or models, schedulers and optimizers are taken
directly form PyTorch. The reason for that is that they are not
typically implemented by anyone else but rather we use the
vanilla implementations that are available.