feat: add generator subpackage with an example
The last subpackage to complete the restructuring has been added:
the generator; together with one of its plausible components (a
density map generator). The whole subpackage will be much more
complex and also generate data in a format the learning part wants
but for now it serves as the structure.

A sample script in the examples folder has been added which makes
use of the simulated data by the simulator to create a heatmap.

See: #9