fix: various fixes to answer Vanessa's comments
`request_device.py` no longer accepts a `self` parameter since
it does not belong to any class. This needed a bit of refactoring
on those parts which were using it to provide a logger to the
function.

The examples folder now contains a data folder that is kept.

The generator example now issues a warning message if the
initial population file is not present. It should be generated
with the simulator.

See: #9 #19