feat: add Hydra for parameter handling in generator example
This is the first approach to handle parameter variability. Instead
of having fixed constants or propagating all of them through function
calls, they have been centralized in a configuration dictionary for
the simulator.

Moreover, such configuration default values can be overridden by
using Hydra and its configuration file or CLI arguments.

Documentation to come in further commits and extension to more
parts of the code.

See: #22