feat: add constants module to simulator
In order to recycle all the constants that are repeated throughout
the code and also sort them in a more logical way, we have gathered
them all together into a constants module.

In the future, we could let this module load the constants from
a JSON file maybe and if we start having too many constants we will
look into ways of grouping them.

See: #21