refactor: automatic channel detection and loading in multi-channel loaders
Before, we had to manually specify the indices for each one of the columns
representing input channels for a sample of the dataset so that the loader
was able to load them. This was inconvenient because we will probably be
changing the amount of columns and inputs a lot so it should be dynamic.

Now, the headers of the dataset CSV files are tagged with an "input:"
prefix to identifiy input columns and distinguish them from the rest
(which are processed as labels or ground truth).

Using this information, the multi-channel loaders now automatically
loop through all columns and detect the input channels to stack them
together to create a multi-channel input.

See PR#39.