feat: add input ignoring to dataset loaders
Now dataset loaders are able to ignore some input columns by dropping
them out of the dataframe loaded from the CSV file. This will provide
us with flexibility when training.

The experiment configuration file has a new parameter that allows us
to indicate a list of channels to ignore. This can be passed as an
argument to the training script as well.

Some other cosmetic changes have been performed.

To reflect this situation, we have also added the default dataset
with 8 samples, 64x64 arrays with position and velocity maps. The
default configuration file has been updated accordingly. The purpose
of this inclusion is to allow people to test things quickly with
a small dataset that we share.

See: PR#39