refactor: unify dataset generators and parameterize them
Now both image/array dataset generators are merged together in a
single file. The type to generate can be chosen via a CLI argument.
The method to generate them is selected at runtime depending on
such parameter. In order to do that, small changes in the dataset
generator routines were needed to have a common interface.

TODO: When we restructure the generator routines, we should also
refactor them to force a common interface.

See: PR#39