fix: improve behaviour of generator normalization
The behaviour of the normalization flag for the dataset generation
could be a bit misleading due to the way that argparse parses the
boolean values. The argument was not required but whenever you
provided a value it was true.

Now the argument is optional and only if it is present it is true.
No need to pass a value, use `--normalize` to set normalization.