feat: basics for multiparameter estimation
This commit starts to tackle issue #51.

Firstly, the simulator helper adds another paramter to its CLI arguments
to help the sweep: h_mean. Notice that we do not need to modify the
simulator at all thanks to Hydra.

Second, the data loaders are modified to accept a new ignore_labels
argument which works in the same way as ignore_inputs.

We have also added label normalization to [0,1] range in the generator.
Since we will have many parameters and probably most of them will have
wildly different ranges, this is a must to make training easy.

The configuration files for the training have been changed to reflect
these additions and a new one has been separated for multiparameter
estimation.

See: #51.