refactor: improve multiparameter training (quality of life)
Now the multiparameter training shows the loss calculation for each
one of the input parameters to allow tracking how it goes down
individually. Updates and backpropagations are done on the
overall loss not individualized.

The metric tracker has been refactored in order not to have to
specify the metrics that we need to track at construction time
but they can now be specified and created on-the-fly at update.

Configuration file for multiparameter loss updated.

See #54.