refactor: allow missing configuration object creation
Before, if any object requested in the code was missing in the
configuration file (the requested name cannot be found in the
configuration dictionary), the creation will fail miserably,
not allowing optional objects.

Now, if the requested objects are not present in the configuration
dictionary they are just not created and assigned None.

This allows defining optional execution paths such as providing
a validation loader or not, or providing a scheduler or not.

See: #52.