feat: add weight initializers module and integrate into pipeline
This commit adds a new module to the pypopsyn learning part which
contains a bunch of weight initializers for a neural network composed
by linear layers (to be extended to convolutional ones later).

It also integrates such initialization process in the training
pipeline (in the examples learning train.py script). Also changes
the configuration file to make use of such initialization scheme.

TODO:
* merge configuration files.
* document initializers.
* experiment to check their behaviors.

See: #58