feat: WIP on-the-fly normalization and standardization
This commit is the beginning for an on-the-fly implementation of both
normalization and standardization procedures.

The idea is to get rid of such processes in the dataset generator and
instead perform that kind of preprocessing on the fly inside the
data loader and the training pipeline.

It is not a simple task for various reasons:
* Normalization is fairly easy and can be done in a per-sample basis.
* Standardization is a bit trickier and should be in a per-dataset basis.
* In order to recover the proper validation metric, the output and
targets have to be de-normalized or de-standardized. We need a more
elegant way to do so.

See: #59