refactor: documentation for normalization, standardization and QoL improvements
A ton of documentation has been added to the loaders to reflect the new
normalization/standardization on-the-fly procedure.

Furthermore, documentation has also been added and code has been reformatted
in the affected sections (trainer mainly and training example). Not only
about this issue but also about previous changes to increase readability.

We have also refactored the custom CLI arguments processing to allow for
a single argument to affect various components by specifying a list separated
by commas in the CustomArgs type.

See: 59