feat: add threshold-based automatic experiment re-running
This commit adds the functionality to the training example to
automatically re-run an experiment as many times as specified
as long as a convergence criteria is not met.

This convergence threshold can be specified either in the
configuration file or via custom CLI arguments.

The training accuracy according to the specified metric (or
the validation one if present) is compared against the
convergence threshold to decide whether or not the
experiment should run again.

Minor changes to the output and a bit of refactoring is
also included in this commit.

See: #62