fix: numba issue with galactic model and dynamical evolution
This commit fixes the numba issue with the dynamical evolution
due to the refactoring of the potential method into classes when
it was independent before. Now the galactic model classes need
to be @jitclassed.

The galactic model election has also been refactored so that
it is initialized once in the very same galactic model file
and held as a global variable accessible by any file to avoid
repetitions of many ifs over here and there.

Furthermore, a base class has been added for the galactic model
to enforce the implementation of needed methods and also to
be able to provide it as a general type for any function.

See: #63