refactor: randomly initialize the NumPy seed and allow overriding it
Now the NumPy random seed is initialized to the current time (a random
way to initialize it) for each simulation run on the initial population
object. Optionally, the user can override the seed by providing a
specific one for reproducibility.

Since the NumPy seed is initialized globally and across imports, we
have also removed all other seed reinitializations throughout the
code to make things simpler and robust (future multithreading?).

TODO: Ideally, this will be done passing a RandomState but it is
a bit of an overkill now.