feat: add benchmarking utilities and make use of them
We have added three different ways of benchmarking: user-defined
scopes with `timewith` and individual checkpoints inside such
scopes, decorators for timing specific functions with `timefunc`,
and decorators for deep profiling of functions with profile output
to files with `cprofile.do_profile`. We will document this more
in the docs.

We have also added working examples of the three situations in the
`initialize_evolve_population.py` example script for the simulator.

Documentation for code and use coming in the next commits.