WIP: timing for backward and forward passes
Added a couple of functions to measure the time of a forward and
a backward pass for a model plus another one to do a set of dry
runs and average the timings of an actual run.

TODO: Pipe fake inputs and outputs with proper input and output
sizes from the data loader.