refactor: remove Hydra depdency from simulator sample script
We no longer use Hydra for the simulator script since it will be
superseeded by an in-house multi-threaded (and later multinode)
implementation.

This commit removes the dependency and makes the simulation receive
two new parameters: an output directory and a path to a JSON file
containing overrides for the configuration parameters.

Both can be provided via CLI.

See: #76